+++ src/imio/helpers/interfaces.py
@@ -2,6 +2,7 @@
 """Module where all interfaces, events and exceptions live."""
 
 from zope.interface import Interface
+from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 
 
 class IListContainedDexterityObjectsForDisplay(Interface):
@@ -12,3 +13,7 @@ class IListContainedDexterityObjectsForDisplay(Interface):
 class IContainerOfUnindexedElementsMarker(Interface):
     """Marker interface to be used to mark a container that contains unindexed elements,
        so elements that are not indexed in the portal_catalog."""
+
+
+class IIMIOHelpersLayer(IDefaultBrowserLayer):
+    """Marker interface that defines a browser layer."""